backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2024.1.1
python-dotenv>=1.0.1
drf-nested-routers>=0.93.5
Pillow>=10.0.0